Splitting tests into diff files
