+++ tests/unit/test_firebolt_dialect.py
@@ -1,22 +1,12 @@
 import os
 from unittest import mock
 
-import pytest
 import sqlalchemy
-from conftest import MockAsyncDBApi, MockDBApi
+from conftest import MockDBApi
 from sqlalchemy.engine import url
-from sqlalchemy.util import await_only, greenlet_spawn
+from sqlalchemy.sql import text
 
 import firebolt_db  # SQLAlchemy package
-from firebolt_db.firebolt_async_dialect import (
-    AsyncAPIWrapper,
-    AsyncConnectionWrapper,
-    AsyncCursorWrapper,
-    AsyncFireboltDialect,
-)
-from firebolt_db.firebolt_async_dialect import (
-    dialect as async_dialect_definition,
-)
 from firebolt_db.firebolt_dialect import (
     FireboltCompiler,
     FireboltDialect,
@@ -24,139 +14,6 @@ from firebolt_db.firebolt_dialect import (
     FireboltTypeCompiler,
 )
 from firebolt_db.firebolt_dialect import dialect as dialect_definition
-
-
-class TestAsyncFireboltDialect:
-    def test_create_dialect(self, async_dialect: AsyncFireboltDialect):
-        assert issubclass(async_dialect_definition, AsyncFireboltDialect)
-        assert type(AsyncFireboltDialect.dbapi()) == AsyncAPIWrapper
-        assert async_dialect.name == "firebolt"
-        assert async_dialect.driver == "firebolt_aio"
-        assert issubclass(async_dialect.preparer, FireboltIdentifierPreparer)
-        assert issubclass(async_dialect.statement_compiler, FireboltCompiler)
-        # SQLAlchemy's DefaultDialect creates an instance of
-        # type_compiler behind the scenes
-        assert isinstance(async_dialect.type_compiler, FireboltTypeCompiler)
-        assert async_dialect.context == {}
-
-    @pytest.mark.asyncio
-    async def test_create_wrapper(self, async_api: mock.AsyncMock(spec=MockAsyncDBApi)):
-        def test_wrapper():
-            async_api.paramstyle = "quoted"
-            wrapper = AsyncAPIWrapper(async_api)
-            assert wrapper.dbapi == async_api
-            assert wrapper.paramstyle == "quoted"
-            wrapper.connect("test arg")
-
-        await greenlet_spawn(test_wrapper)
-        async_api.connect.assert_called_once_with("test arg")
-
-    @pytest.mark.asyncio
-    async def test_connection_wrapper(
-        self, async_api: mock.AsyncMock(spec=MockAsyncDBApi)
-    ):
-        def test_connection():
-            wrapper = AsyncConnectionWrapper(async_api, await_only(async_api.connect()))
-            # Check call propagation
-            wrapper.commit()
-            wrapper.rollback()
-            wrapper.close()
-            assert isinstance(wrapper.cursor(), AsyncCursorWrapper)
-
-        await greenlet_spawn(test_connection)
-        async_api.connect.return_value.commit.assert_awaited_once()
-        async_api.connect.return_value.rollback.assert_awaited_once()
-        async_api.connect.return_value._aclose.assert_awaited_once()
-
-    @pytest.mark.asyncio
-    async def test_cursor_execute(
-        self,
-        async_api: mock.AsyncMock(spec=MockAsyncDBApi),
-        async_connection,
-        async_cursor,
-    ):
-        def test_cursor():
-            async_connection.cursor.return_value = async_cursor
-            conn_wrapper = AsyncConnectionWrapper(async_api, async_connection)
-            wrapper = AsyncCursorWrapper(conn_wrapper)
-            wrapper.execute("INSERT INTO test(a, b) VALUES (?, ?)", [(1, "a")])
-            assert wrapper.description == async_cursor.description
-            assert wrapper.rowcount == async_cursor.rowcount
-
-        async_cursor.rowcount = -1
-        await greenlet_spawn(test_cursor)
-        async_cursor.execute.assert_awaited_once_with(
-            "INSERT INTO test(a, b) VALUES (?, ?)", [(1, "a")]
-        )
-        async_cursor.fetchall.assert_awaited_once()
-        async_cursor.close.assert_called_once()
-
-        async_cursor.execute.reset_mock()
-        async_cursor.fetchall.reset_mock()
-        async_cursor.close.reset_mock()
-        async_cursor.description = None
-        async_cursor.rowcount = 100
-
-        await greenlet_spawn(test_cursor)
-        async_cursor.execute.assert_awaited_once_with(
-            "INSERT INTO test(a, b) VALUES (?, ?)", [(1, "a")]
-        )
-        async_cursor.fetchall.assert_not_awaited()
-        async_cursor.close.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_cursor_close(
-        self,
-        async_api: mock.AsyncMock(spec=MockAsyncDBApi),
-        async_connection,
-        async_cursor,
-    ):
-        def test_cursor():
-            async_connection.cursor.return_value = async_cursor
-            conn_wrapper = AsyncConnectionWrapper(async_api, async_connection)
-            wrapper = AsyncCursorWrapper(conn_wrapper)
-            wrapper._rows = [1, 2, 3]
-            wrapper.close()
-            assert wrapper._rows == []
-
-        await greenlet_spawn(test_cursor)
-
-    @pytest.mark.asyncio
-    async def test_cursor_executemany(
-        self,
-        async_api: mock.AsyncMock(spec=MockAsyncDBApi),
-        async_connection,
-        async_cursor,
-    ):
-        def test_cursor():
-            async_connection.cursor.return_value = async_cursor
-            conn_wrapper = AsyncConnectionWrapper(async_api, async_connection)
-            wrapper = AsyncCursorWrapper(conn_wrapper)
-            with pytest.raises(NotImplementedError):
-                wrapper.executemany(
-                    "INSERT INTO test(a, b) VALUES (?, ?)", [(1, "a"), (2, "b")]
-                )
-
-        await greenlet_spawn(test_cursor)
-
-    @pytest.mark.asyncio
-    async def test_cursor_fetch(
-        self,
-        async_api: mock.AsyncMock(spec=MockAsyncDBApi),
-        async_connection,
-        async_cursor,
-    ):
-        def test_cursor():
-            async_connection.cursor.return_value = async_cursor
-            conn_wrapper = AsyncConnectionWrapper(async_api, async_connection)
-            wrapper = AsyncCursorWrapper(conn_wrapper)
-            wrapper._rows = [1, 2, 3, 4, 5, 6]
-            assert wrapper.fetchone() == 1
-            assert wrapper.fetchmany() == [2]
-            assert wrapper.fetchmany(2) == [3, 4]
-            assert wrapper.fetchall() == [5, 6]
-
-        await greenlet_spawn(test_cursor)
 
 
 class TestFireboltDialect:
@@ -203,8 +60,9 @@ class TestFireboltDialect:
         ]
         result = dialect.get_schema_names(connection)
         assert result == ["schema1", "schema2"]
-        connection.execute.assert_called_once_with(
-            "select schema_name from information_schema.databases"
+        connection.execute.assert_called_once()
+        assert str(connection.execute.call_args[0][0].compile()) == str(
+            text("select schema_name from information_schema.databases").compile()
         )
 
     def test_table_names(
@@ -220,15 +78,19 @@ class TestFireboltDialect:
 
         result = dialect.get_table_names(connection)
         assert result == ["table1", "table2"]
-        connection.execute.assert_called_once_with(
-            "select table_name from information_schema.tables"
+        connection.execute.assert_called_once()
+        assert str(connection.execute.call_args[0][0].compile()) == str(
+            text("select table_name from information_schema.tables").compile()
         )
         connection.execute.reset_mock()
         result = dialect.get_table_names(connection, schema="schema")
         assert result == ["table1", "table2"]
-        connection.execute.assert_called_once_with(
-            "select table_name from information_schema.tables"
-            " where table_schema = 'schema'"
+        connection.execute.assert_called_once()
+        assert str(connection.execute.call_args[0][0].compile()) == str(
+            text(
+                "select table_name from information_schema.tables"
+                " where table_schema = 'schema'"
+            ).compile()
         )
 
     def test_view_names(
@@ -289,7 +151,10 @@ class TestFireboltDialect:
                     "default": None,
                 },
             ]
-            connection.execute.assert_called_once_with(expected_query)
+            connection.execute.assert_called_once()
+            assert str(connection.execute.call_args[0][0].compile()) == str(
+                text(expected_query).compile()
+            )
             connection.execute.reset_mock()
 
     def test_pk_constraint(
